Updated recurring job display
Removed the save button from the admin header actions in the RecurringJob view. Replaced job ID with a new DisplayName property in the table. Added this DisplayName property to the RecurringJobInfo model and updated it in both active and inactive jobs within the RecurringJobAdminService. Also, removed hardcoded Ids from various jobs and replaced them with dynamic values based on their type names, which are also used as DisplayNames.