Fix several issues regarding new changes
Refactor RefreshTextSearchIndex class and move to Jobs namespace

The `RefreshTextSearchIndex` class was refactored and moved from the `Services` namespace to the `Jobs` namespace. The class was also renamed to `RefreshTextSearchIndexTask`. This change improves code organization by grouping job-related classes together. Additionally, the ISession dependency was removed from the constructor as it was not being used, simplifying the codebase.