Improved navigation in admin layout area
The changes made improve the user experience in the admin layout area by ensuring that after adding or deleting a widget, users are redirected back to their previous page. This was achieved by including 'returnUrl' parameter in the URL actions for Add and Delete operations. The unnecessary line of code was also removed from the form section for cleaner code structure.